requirements.txt
Flask
Jinja2
psycopg2-binary
pika
